+++ crates/gpui/Cargo.toml
@@ -81,6 +81,7 @@ doctest = false
 anyhow.workspace = true
 async-task = "4.7"
 backtrace = { version = "0.3", optional = true }
+bevy_utils_proc_macros = "0.15.2"
 blade-graphics = { workspace = true, optional = true }
 blade-macros = { workspace = true, optional = true }
 blade-util = { workspace = true, optional = true }
