+++ crates/gpui/src/app.rs
@@ -25,6 +25,7 @@ use collections::{FxHashMap, FxHashSet, HashMap, VecDeque};
 pub use context::*;
 pub use entity_map::*;
 use http_client::{HttpClient, Url};
+pub use plugin::*;
 use smallvec::SmallVec;
 #[cfg(any(test, feature = "test-support"))]
 pub use test_context::*;
@@ -45,9 +46,12 @@ use crate::{
     current_platform, hash, init_app_menus,
 };
 
+pub use self::plugin::InsertGlobal;
+
 mod async_context;
 mod context;
 mod entity_map;
+mod plugin;
 #[cfg(any(test, feature = "test-support"))]
 mod test_context;
 
@@ -168,17 +172,24 @@ impl Application {
         self
     }
 
+    /// Adds a plugin or group of plugins to the appliication
+    pub fn add_plugins<M>(mut self, plugins: impl Plugins<M>) -> Self {
+        let mut context_lock = self.0.borrow_mut();
+        plugins.add_to_app(&mut context_lock.0);
+        drop(context_lock);
+        self
+    }
+
     /// Start the application. The provided callback will be called once the
     /// app is fully launched.
-    pub fn run<F>(self, on_finish_launching: F)
-    where
-        F: 'static + FnOnce(&mut App),
-    {
+    pub fn run(self) {
         let this = self.0.clone();
         let platform = self.0.borrow().platform.clone();
         platform.run(Box::new(move || {
             let cx = &mut *this.borrow_mut();
-            on_finish_launching(cx);
+            while let Some(plugin) = cx.plugins.pop_front() {
+                plugin.build(cx);
+            }
         }));
     }
 
@@ -291,6 +302,7 @@ pub struct App {
     #[cfg(any(test, feature = "test-support", debug_assertions))]
     pub(crate) name: Option<&'static str>,
     quitting: bool,
+    pub(crate) plugins: VecDeque<Box<dyn Plugin>>,
 }
 
 impl App {
@@ -361,6 +373,7 @@ impl App {
 
                 #[cfg(any(test, feature = "test-support", debug_assertions))]
                 name: None,
+                plugins: Default::default(),
             }),
         });
 
@@ -387,6 +400,19 @@ impl App {
         }));
 
         app
+    }
+
+    /// Adds the given plugin or plugins to the application.
+    pub fn add_plugins<M>(&mut self, plugins: impl Plugins<M>) -> &mut Self {
+        plugins.add_to_app(self);
+        self
+    }
+
+    /// Assign the given asset source to the application.
+    pub fn with_assets(&mut self, asset_source: impl AssetSource) {
+        let asset_source = Arc::new(asset_source);
+        self.asset_source = asset_source.clone();
+        self.svg_renderer = SvgRenderer::new(asset_source);
     }
 
     /// Quit the application gracefully. Handlers registered with [`Context::on_app_quit`]
@@ -1383,6 +1409,31 @@ impl App {
                     listener(action, cx)
                 }
             }));
+    }
+
+    /// Register a handler to be invoked when the platform instructs the application
+    /// to open one or more URLs.
+    pub fn on_open_urls<F>(&self, mut callback: F) -> &Self
+    where
+        F: 'static + FnMut(Vec<String>),
+    {
+        self.platform.on_open_urls(Box::new(callback));
+        self
+    }
+
+    /// Invokes a handler when an already-running application is launched.
+    /// On macOS, this can occur when the application icon is double-clicked or the app is launched via the dock.
+    pub fn on_reopen<F>(&self, mut callback: F) -> &Self
+    where
+        F: 'static + FnMut(&mut App),
+    {
+        let this = self.this.clone();
+        self.platform.on_reopen(Box::new(move || {
+            if let Some(app) = this.upgrade() {
+                callback(&mut app.borrow_mut());
+            }
+        }));
+        self
     }
 
     /// Event handlers propagate events by default. Call this method to stop dispatching to
