+++ crates/zed/Cargo.toml
@@ -10,14 +10,6 @@ authors = ["Zed Team <hi@zed.dev>"]
 [lints]
 workspace = true
 
-[[bin]]
-name = "zed"
-path = "src/zed-main.rs"
-
-[lib]
-name = "zed"
-path = "src/main.rs"
-
 [dependencies]
 activity_indicator.workspace = true
 agent.workspace = true
